Split oversized pool and dashboard modules
