+++ indexer-envio/src/pool.ts
@@ -1,408 +1,65 @@
 // ---------------------------------------------------------------------------
-// Pool and PoolSnapshot upsert logic, health status computation
+// Pool upsert logic and public pool-module re-exports
 // ---------------------------------------------------------------------------
 
-import type { EffectCaller } from "envio";
-import type {
-  Pool,
-  PoolSnapshot,
-  PoolDailySnapshot,
-  DeviationThresholdBreach,
-} from "generated";
-import type { HandlerContext } from "generated/src/Types";
-import { ZERO_ADDRESS } from "./constants";
-import {
-  hourBucket,
-  dayBucket,
-  snapshotId,
-  dailySnapshotId,
-  extractAddressFromPoolId,
-} from "./helpers";
-import {
-  computePriceDifference,
-  parseDecimalsPair,
-  scalingFactorToDecimals,
-} from "./priceDifference";
+import type { Pool } from "generated";
+import { extractAddressFromPoolId, isVirtualPool } from "./helpers";
+import { computePriceDifference } from "./priceDifference";
 import {
   compactFees,
   feesEffect,
-  invertRateFeedEffect,
-  poolExchangeEffect,
-  rebalanceThresholdsEffect,
   referenceRateFeedIDEffect,
   reportExpiryEffect,
-  tokenDecimalsScalingEffect,
-  vpExchangeIdEffect,
 } from "./rpc/effects";
-import { lookupPricingModuleName } from "./contractAddresses";
-import { isVirtualPool } from "./helpers";
 import { recordBreachTransition } from "./deviationBreach";
+import {
+  computeHealthStatus,
+  isNeverRebalance,
+  nextDeviationBreachStartedAt,
+  nextOpenBreachEntryThreshold,
+  nextOpenBreachPeak,
+} from "./pool/health";
+import { pickPreferredSource, type PoolUpdateSource } from "./pool/sources";
+import {
+  selfHealInvertRateFeed,
+  selfHealTokenDecimals,
+  selfHealWrappedExchangeId,
+} from "./pool/self-heal";
+import type { PoolContext } from "./pool/types";
 
-// ---------------------------------------------------------------------------
-// Health status computation
-// ---------------------------------------------------------------------------
-
-/**
- * How long a pool may sit above the critical magnitude (5% over threshold)
- * before the status escalates from WARN to CRITICAL. Mirrors
- * `DEVIATION_BREACH_GRACE_SECONDS` in `ui-dashboard/src/lib/health.ts`.
- */
-export const DEVIATION_BREACH_GRACE_SECONDS = 3600n;
-
-/**
- * Tolerance + critical-magnitude thresholds as `num/den` pairs over the
- * rebalance threshold. Integer math avoids float pathology at the boundaries.
- *
- * Float-form mirrors live in `@mento-protocol/monitoring-config/thresholds`
- * (canonical for the dashboard + metrics-bridge probe). Parity with the
- * dashboard's float comparison is enforced by `test/healthStatusParity.test.ts`.
- * Any change here must update that file too.
- */
-export const DEVIATION_TOLERANCE_NUM = 101n;
-export const DEVIATION_TOLERANCE_DEN = 100n;
-export const DEVIATION_CRITICAL_NUM = 105n;
-export const DEVIATION_CRITICAL_DEN = 100n;
-
-/**
- * Health-status union the indexer can emit. Narrower than the dashboard's
- * `HealthStatus` (no "WEEKEND" — that's a render-time reclassification of
- * stale-oracle CRITICAL).
- */
-export type IndexerHealthStatus = "OK" | "WARN" | "CRITICAL" | "N/A";
-
-/** True iff governance has explicitly configured this pool to never
- * rebalance — BOTH `rebalanceThresholdAbove === 0` AND
- * `rebalanceThresholdBelow === 0` AND `rebalanceThresholdsKnown=true`.
- * Distinct from the schema-default unknown case (`rebalanceThresholdsKnown=false`),
- * where the breach predicate falls back to a 10000-bps under-bound until
- * self-heal lands.
- *
- * Cannot infer from `rebalanceThreshold` alone: that's the ACTIVE side
- * picked by `pickActiveThreshold` based on current reserve direction. An
- * asymmetric pool with `above=0, below=300` legitimately persists
- * `rebalanceThreshold=0` while reservePrice is on the above side, even
- * though the pool DOES rebalance on the below side. Both split fields
- * must be 0 for "never rebalance" to be the right semantic.
- *
- * Used to short-circuit breach/health predicates rather than relying on
- * the `effectiveThreshold` 1e12 sentinel: explicit short-circuit means an
- * extreme reserve-skew priceDifference > 1.01e12 still resolves to "no
- * breach", as governance intended.
- */
-export const isNeverRebalance = (pool: {
-  rebalanceThresholdAbove?: number;
-  rebalanceThresholdBelow?: number;
-  rebalanceThresholdsKnown?: boolean;
-}): boolean =>
-  // STRICT equality: undefined is NOT treated as 0. A pool entity always
-  // has both split fields populated (Int! schema), so this only matters
-  // for synthetic test inputs. Defaulting undefined→0 would let a caller
-  // claim never-rebalance without populating the split fields, which is
-  // the same partial-shape pitfall the dashboard mirror guards against.
-  pool.rebalanceThresholdAbove === 0 &&
-  pool.rebalanceThresholdBelow === 0 &&
-  pool.rebalanceThresholdsKnown === true;
-
-/** Resolve the effective threshold in bps. Three states:
- *  - `> 0`: the on-chain configured threshold (active side).
- *  - `0` AND `rebalanceThresholdsKnown=true`: governance configured the pool
- *    to never rebalance. Treat as effectively infinite so the breach
- *    predicate never trips — the pool isn't supposed to rebalance. Callers
- *    that take a different code path on never-rebalance pools should also
- *    check `isNeverRebalance(pool)` and short-circuit upstream rather than
- *    relying on the 1e12 sentinel — that cushion is unbounded-skew-tolerant
- *    in practice but not by construction.
- *  - `0` AND `rebalanceThresholdsKnown=false`: indexer hasn't read on-chain
- *    yet. Fall back to 10000 (100%) so the predicate doesn't false-trip
- *    while waiting for self-heal.
- *
- * The schema-default unknown case must NOT collapse with the legitimate
- * "never rebalance" case — both have `rebalanceThreshold === 0`, but only
- * the latter has `rebalanceThresholdsKnown=true`.
- */
-export const effectiveThreshold = (pool: {
-  rebalanceThreshold: number;
-  rebalanceThresholdAbove?: number;
-  rebalanceThresholdBelow?: number;
-  // Optional: callers passing a synthetic threshold value (e.g.
-  // `deviationBreach` healing from a captured entry threshold) may not
-  // carry the Known flag. In that case we treat `0` as the unread
-  // sentinel and fall back to 10000 — matches pre-Lever-4 behaviour.
-  rebalanceThresholdsKnown?: boolean;
-}): bigint => {
-  if (pool.rebalanceThreshold > 0) return BigInt(pool.rebalanceThreshold);
-  // Distinguish governance-disabled "never rebalance" (BOTH split sides
-  // 0 + Known) from an asymmetric pool whose active side just happens to
-  // be 0 right now. Otherwise an `above=0, below=300` pool with reserves
-  // currently picking the above side would suppress all deviation
-  // alerts via the 1e12 cushion even though the below side is real.
-  if (isNeverRebalance(pool)) return 10n ** 12n;
-  return 10000n;
-};
-
-/**
- * Persistable counterpart to `effectiveThreshold` for `Int!`-typed fields
- * (GraphQL 32-bit signed, max ~2.1e9). Three distinct zero-cases need
- * three distinct persisted values so chart/table consumers don't have to
- * synthesize them from joins:
- *
- * - **never-rebalance** (`above=0, below=0, known=true`): persist `0`.
- *   Matches on-chain config; `isNeverRebalance(pool)` short-circuits the
- *   chart/health math upstream, so the value is informational. The 1e12
- *   in-memory sentinel would overflow `Int!` here.
- * - **asymmetric-active-zero** (one split side `>0`, other `=0`,
- *   `known=true`, reserves currently picking the zero side): persist
- *   `10000` (the same fallback `effectiveThreshold` uses for breach
- *   scoring). Without this, `oracle-chart.tsx` renders deviation as `0%`
- *   and the table renders `—` even though `hasHealthData=true` and the
- *   breach predicate scored against 10000 — making valid breach samples
- *   look in-band.
- * - **unknown-zero** (`known=false`, raw threshold `=0` because indexer
- *   hasn't read on-chain): persist `0`. The row's `hasHealthData=false`
- *   flag is what gates consumers; persisting 10000 here would leak the
- *   fallback into rows that the indexer explicitly marked untrusted.
- *
- * For any non-zero raw threshold, return it directly (already `Int!`-safe).
- */
-export const persistableThreshold = (pool: {
-  rebalanceThreshold: number;
-  rebalanceThresholdAbove?: number;
-  rebalanceThresholdBelow?: number;
-  rebalanceThresholdsKnown?: boolean;
-}): number => {
-  if (pool.rebalanceThreshold > 0) return pool.rebalanceThreshold;
-  if (isNeverRebalance(pool)) return 0;
-  // From here: raw = 0 AND not never-rebalance. If thresholds are known,
-  // this is the asymmetric-active-zero case (the OTHER side is positive
-  // and the pool DOES rebalance on flip) — persist the 10000 fallback so
-  // the chart's deviation-ratio math has a non-zero denominator. If
-  // thresholds are unread, persist 0; `hasHealthData=false` on the row
-  // signals consumers to skip.
-  if (pool.rebalanceThresholdsKnown === true) return 10000;
-  return 0;
-};
-
-/**
- * Breach-row entry capture: the threshold the breach predicate
- * (`isInDeviationBreach` → `effectiveThreshold(pool)`) was scored against
- * at the rising edge. Differs from `persistableThreshold` because the
- * breach row's `entryRebalanceThreshold` field exists to score severity
- * across the breach lifecycle — capturing raw 0 on an asymmetric pool's
- * zero-threshold side would let a later reserve flip re-score history
- * against the post-flip opposite side (codex P2 #3214513401, PR 1.6).
- *
- * Returns:
- *  - active threshold (bps) when positive — symmetric / on-active-side asymmetric.
- *  - 10000 when active is 0 AND `rebalanceThresholdsKnown=false` (cold-start
- *    under-bound — predicate scored against the same fallback) OR the pool
- *    is asymmetric on its zero side (`above=0, below>0` etc.).
- *  - 0 when `isNeverRebalance` — returning the never-rebalance 1e12 cushion
- *    would overflow `Int!`. Never-rebalance pools cannot have an open breach
- *    (the predicate short-circuits them), so this branch is defense-in-depth
- *    only — if reached, the row's later `criticalDurationSeconds` accrual
- *    would resolve to 0 via `effectiveThreshold(pool)`'s 1e12 cushion.
- */
-export const breachEntryThreshold = (pool: {
-  rebalanceThreshold: number;
-  rebalanceThresholdAbove?: number;
-  rebalanceThresholdBelow?: number;
-  rebalanceThresholdsKnown?: boolean;
-}): number => {
-  if (isNeverRebalance(pool)) return 0;
-  return Number(effectiveThreshold(pool));
-};
-
-/** True when `priceDifference` is strictly above the 5% critical-magnitude
- * line, integer-safe. Used by both the live status branch (here) and the
- * cumulative `criticalDurationSeconds` accrual in `deviationBreach.ts` to
- * keep them in lockstep. */
-export const isAboveCriticalMagnitude = (
-  priceDifference: bigint,
-  threshold: bigint,
-): boolean =>
-  priceDifference * DEVIATION_CRITICAL_DEN > threshold * DEVIATION_CRITICAL_NUM;
-
-/**
- * Mirror of `computeHealthStatus` in `ui-dashboard/src/lib/health.ts`; parity
- * is enforced by `test/healthStatusParity.test.ts`. The breach anchor
- * (`deviationBreachStartedAt`) is set at the 1.01x crossing in
- * `isInDeviationBreach`, so the 1h grace counts from when the pool first
- * exceeded tolerance.
- *
- * Intentional divergences NOT covered by the parity suite:
- *  - Oracle staleness: indexer reads the event-time `oracleOk` flag; the UI
- *    reads `oracleTimestamp + oracleExpiry` against wall clock at render time
- *    with per-chain fallbacks.
- *  - Weekend reclassification: only the UI has `isWeekend()` at render time.
- *    Indexed weekend-stale pools surface as CRITICAL here; the UI
- *    reclassifies them to WEEKEND.
- *  - `hasHealthData=false` short-circuit: defense-in-depth mirror of the
- *    UI's gate at `health.ts:230`. Indexer-side, the upstream
- *    `tokenDecimalsKnown` early-return in sortedOracles handlers prevents
- *    `computeHealthStatus` from being called on untrusted samples in the
- *    first place — but mirroring the gate keeps the function pure and
- *    answers any caller (parity tests, ad-hoc replay tooling) that
- *    bypasses the upstream guard.
- */
-export function computeHealthStatus(
-  pool: Pool,
-  nowSeconds: bigint,
-): IndexerHealthStatus {
-  if (isVirtualPool(pool)) return "N/A";
-  // `oracleOk=false` is an alertable freshness incident — keep it ABOVE the
-  // hasHealthData gate so a stale-oracle pool doesn't get masked into "N/A"
-  // just because the deviation accrual is also untrusted (codex P2 PR #370
-  // #3214756056).
-  if (!pool.oracleOk) return "CRITICAL";
-  if (pool.hasHealthData === false) return "N/A";
-  // Governance-configured "never rebalance" pools stay OK regardless of
-  // priceDifference magnitude. Short-circuit explicitly so extreme reserve
-  // skew can't trip the predicate via the `effectiveThreshold` 1e12 cushion.
-  if (isNeverRebalance(pool)) return "OK";
-  const threshold = effectiveThreshold(pool);
-  const diff = pool.priceDifference;
-  const aboveTolerance =
-    diff * DEVIATION_TOLERANCE_DEN > threshold * DEVIATION_TOLERANCE_NUM;
-  if (!aboveTolerance) return "OK";
-  if (!isAboveCriticalMagnitude(diff, threshold)) return "WARN";
-  // Without a breach-start anchor (indexer hasn't populated it yet), stay
-  // at WARN rather than spuriously escalating to CRITICAL.
-  if (pool.deviationBreachStartedAt <= 0n) return "WARN";
-  const withinGrace =
-    nowSeconds - pool.deviationBreachStartedAt < DEVIATION_BREACH_GRACE_SECONDS;
-  return withinGrace ? "WARN" : "CRITICAL";
-}
-
-// Strict `>` at the tolerance line matches `computeHealthStatus`. Oracle
-// staleness is intentionally NOT counted — this tracks price action only.
-// Never-rebalance pools always short-circuit to false (mirrors
-// `computeHealthStatus`); see `isNeverRebalance` for why.
-export function isInDeviationBreach(pool: Pool): boolean {
-  if (isVirtualPool(pool)) return false;
-  if (isNeverRebalance(pool)) return false;
-  return (
-    pool.priceDifference * DEVIATION_TOLERANCE_DEN >
-    effectiveThreshold(pool) * DEVIATION_TOLERANCE_NUM
-  );
-}
-
-export function nextDeviationBreachStartedAt(
-  prev: Pool | undefined,
-  next: Pool,
-  blockTimestamp: bigint,
-  source?: PoolUpdateSource,
-): bigint {
-  const wasBreachedPrice = prev ? isInDeviationBreach(prev) : false;
-  const wasBreachedAnchor = prev ? prev.deviationBreachStartedAt > 0n : false;
-  const isBreached = isInDeviationBreach(next);
-  if (!isBreached) {
-    // Defer the close when this transition is being driven by
-    // UpdateReserves. The FPMM contract emits ReservesUpdated inside
-    // swap/rebalance/mint/burn (often MULTIPLE times — pre- and post-
-    // state), so an initial UR can pull priceDifference to / below
-    // threshold before the semantic handler runs. Use the ANCHOR, not
-    // price, to decide "is there an open breach to hold" — price may
-    // already read healthy after UR#1, but the anchor is still set.
-    // Holding it keeps the falling-edge attribution with the eventual
-    // semantic handler (Rebalance → "rebalance", Swap → "swap", etc.)
-    // instead of the generic UR "unknown".
-    if (wasBreachedAnchor && source === "fpmm_update_reserves" && prev) {
-      return prev.deviationBreachStartedAt;
-    }
-    return 0n;
-  }
-  if (!wasBreachedPrice) return blockTimestamp;
-  // Self-heal: a breached row with a 0n sentinel (partial restore, pre-backfill
-  // state, etc) would stay 0n forever. Adopt the current block time as a
-  // best-effort start so the UI stops suppressing the indicator.
-  return prev!.deviationBreachStartedAt > 0n
-    ? prev!.deviationBreachStartedAt
-    : blockTimestamp;
-}
-
-/** Maintain the open-breach peak denormalized on Pool. Mirrors the
- * `peakPriceDifference` tracked on the open `DeviationThresholdBreach`
- * row, but lives on Pool so the rollup query the live uptime tile uses
- * doesn't need to join to the breach row. Resets to 0 when no open
- * breach; otherwise carries `max(prev peak, current diff)`. */
-export function nextOpenBreachPeak(prev: Pool | undefined, next: Pool): bigint {
-  if (next.deviationBreachStartedAt === 0n) return 0n;
-  const prevPeak = prev?.currentOpenBreachPeak ?? 0n;
-  return prevPeak > next.priceDifference ? prevPeak : next.priceDifference;
-}
-
-/** Maintain the open-breach entry threshold denormalized on Pool. Captures
- * the EFFECTIVE threshold at the rising edge (`breachEntryThreshold(next)`,
- * not raw `rebalanceThreshold`) so the live-uptime gate scores the peak
- * against the same threshold the breach predicate (`isInDeviationBreach` →
- * `effectiveThreshold(pool)`) used. Asymmetric pools on their zero-threshold
- * side need this: raw `rebalanceThreshold === 0` while the predicate scored
- * against the 10000-bps fallback; capturing 0 here would let the closing
- * fallback chain in `recordBreachTransition` reach for the post-flip
- * opposite side's active value instead. Held across continuing breach events
- * so a mid-breach `FPMMRebalanceThresholdUpdated` (or a side flip) can't
- * shift the live verdict. Resets to 0 when no open breach. */
-export function nextOpenBreachEntryThreshold(
-  prev: Pool | undefined,
-  next: Pool,
-): number {
-  if (next.deviationBreachStartedAt === 0n) return 0;
-  const prevAnchor = prev?.deviationBreachStartedAt ?? 0n;
-  // Rising edge — capture the predicate-scoring threshold (10000 fallback
-  // for asymmetric-zero-side pools), matching the entity row capture in
-  // `recordBreachTransition`. See deviationBreach.ts for the full
-  // asymmetric-side-flip rationale (codex P2 #3214513401, PR 1.6).
-  if (prevAnchor === 0n) return breachEntryThreshold(next);
-  // Continuing: never overwrite a captured value. The pre-PR-1.6
-  // heal-from-zero branch is retired for the same reason as the entity
-  // row's heal (see deviationBreach.ts) — overwriting an old captured 0
-  // with `next.rebalanceThreshold` would re-score history against the
-  // post-flip opposite side. Old rows with stored=0 stay 0; the closing
-  // fallback chain in `recordBreachTransition` defaults them to the
-  // 10000 effective floor.
-  return prev?.currentOpenBreachEntryThreshold ?? 0;
-}
+export {
+  DEVIATION_BREACH_GRACE_SECONDS,
+  DEVIATION_CRITICAL_DEN,
+  DEVIATION_CRITICAL_NUM,
+  DEVIATION_TOLERANCE_DEN,
+  DEVIATION_TOLERANCE_NUM,
+  breachEntryThreshold,
+  computeHealthStatus,
+  effectiveThreshold,
+  isAboveCriticalMagnitude,
+  isInDeviationBreach,
+  isNeverRebalance,
+  nextDeviationBreachStartedAt,
+  nextOpenBreachEntryThreshold,
+  nextOpenBreachPeak,
+  persistableThreshold,
+} from "./pool/health";
+export type { IndexerHealthStatus } from "./pool/health";
+export type { PoolUpdateSource } from "./pool/sources";
+export {
+  mirrorFeedIdToPool,
+  mirrorTokensAndDecimalsToPool,
+  selfHealInvertRateFeed,
+  selfHealRebalanceThresholds,
+  selfHealTokenDecimals,
+  selfHealWrappedExchangeId,
+} from "./pool/self-heal";
+export type { PoolContext, SnapshotContext } from "./pool/types";
+export { upsertDailySnapshot, upsertSnapshot } from "./pool/snapshots";
 
 // ---------------------------------------------------------------------------
 // Pool upsert (with cumulative fields)
 // ---------------------------------------------------------------------------
-
-const SOURCE_PRIORITY = {
-  virtual_pool_factory: 100,
-  fpmm_factory: 90,
-  fpmm_rebalanced: 50,
-  fpmm_update_reserves: 40,
-  // Below state-sync events: a threshold update doesn't change reserves
-  // or oracle, so the legacy "preferred-source" stickiness should keep
-  // whichever live event source wrote last.
-  fpmm_threshold_updated: 35,
-  fpmm_swap: 30,
-  fpmm_mint: 20,
-  fpmm_burn: 20,
-} as const;
-
-/** Values the indexer passes as `source` when calling upsertPool / the
- *  breach helpers. Typing this as a union (rather than bare string) means
- *  a typo like "fpmm_update_reseves" is a compile error instead of a
- *  silently-unmatched deferral branch. */
-export type PoolUpdateSource = keyof typeof SOURCE_PRIORITY;
-
-// `existingSource` is typed as `string` because Pool.source is stored as
-// a plain string in the DB (potentially including legacy values not in
-// the current union). Use a safe lookup helper so unknown strings fall
-// through to priority 0 without an unchecked cast.
-const sourcePriority = (source: string): number =>
-  (SOURCE_PRIORITY as Record<string, number>)[source] ?? 0;
-
-const pickPreferredSource = (
-  existingSource: string | undefined,
-  incomingSource: PoolUpdateSource,
-): string => {
-  if (!existingSource) return incomingSource;
-  return sourcePriority(incomingSource) >= sourcePriority(existingSource)
-    ? incomingSource
-    : existingSource;
-};
 
 /**
  * Preload-phase helper used by every event handler that makes direct RPC
@@ -440,553 +97,6 @@ export async function maybePreloadPool(
   );
   return true;
 }
-
-export type PoolContext = {
-  effect: EffectCaller;
-  Pool: {
-    get: (id: string) => Promise<Pool | undefined>;
-    set: (entity: Pool) => void;
-  };
-  DeviationThresholdBreach: {
-    get: (id: string) => Promise<DeviationThresholdBreach | undefined>;
-    set: (entity: DeviationThresholdBreach) => void;
-  };
-  // Used by `selfHealWrappedExchangeId` to patch the back-reference on the
-  // matching exchange row when a VP heals its `wrappedExchangeId`.
-  BiPoolExchange: HandlerContext["BiPoolExchange"];
-};
-
-/** Self-heal `invertRateFeed` when it was never successfully read at pool
- * deployment (factory's RPC fan-out hit a transient blip → the field rode
- * the schema default). Returns the same pool when already healed / not
- * applicable, otherwise returns a copy with `invertRateFeed` corrected and
- * `invertRateFeedKnown: true`.
- *
- * Must be called before any code path reads `pool.invertRateFeed` to
- * compute oracle/health/priceDifference state — including the
- * `OracleReported`/`MedianUpdated` handlers (which write directly without
- * going through `upsertPool`) and the `UpdateReserves`/`Rebalanced`
- * handlers (which read `existing.invertRateFeed` before `upsertPool` runs).
- *
- * Effect-level dedup means this is one RPC read per (pool, batch) when
- * unhealed; once `invertRateFeedKnown` flips true, subsequent calls are
- * pure object identity returns — no RPC, no Pool.set side-effect. The
- * caller's own Pool.set persists the healed value. */
-export async function selfHealInvertRateFeed(
-  context: { effect: EffectCaller },
-  pool: Pool,
-): Promise<Pool> {
-  if (pool.invertRateFeedKnown || pool.source === "" || isVirtualPool(pool)) {
-    return pool;
-  }
-  const poolAddr = extractAddressFromPoolId(pool.id);
-  const invert = await context.effect(invertRateFeedEffect, {
-    chainId: pool.chainId,
-    poolAddress: poolAddr,
-  });
-  if (invert === undefined) return pool;
-  return {
-    ...pool,
-    invertRateFeed: invert,
-    invertRateFeedKnown: true,
-  };
-}
-
-/** Self-heal `token0Decimals` / `token1Decimals` when the factory's
- * `tokenDecimalsScalingEffect` reads failed at deploy. Without this, a
- * non-18-decimal pool whose factory RPC blipped would silently keep the
- * schema default 18/18 forever — `normalizeTo18` would not scale the
- * affected reserve, and every priceDifference computation downstream
- * would be wrong by `10^(18 - real_dec)` factor.
- *
- * Cache-true effect: `tokenDecimalsScalingEffect` is per-(chain, pool, fn)
- * and the on-chain decimals are immutable, so this is one RPC pair per
- * unhealed pool across the entire run. Once `tokenDecimalsKnown` flips
- * true, subsequent calls are pure object-identity returns.
- *
- * Caller's own `Pool.set` persists the healed values. */
-export async function selfHealTokenDecimals(
-  context: { effect: EffectCaller },
-  pool: Pool,
-): Promise<Pool> {
-  // VPs were previously skipped because they don't go through the local
-  // priceDifference recompute path. But `buildSwapTraderFields` still uses
-  // `token0Decimals` / `token1Decimals` to compute `volumeUsdWei` and
-  // leaderboard snapshots — a non-18-decimal USD leg would stay mis-scaled
-  // for a VP with a deploy-time decimal blip. Cache:true effect, so the
-  // RPC pair fires once per VP across the run regardless of how many
-  // events touch it.
-  if (pool.tokenDecimalsKnown || pool.source === "") {
-    return pool;
-  }
-  // Skip when both pair tokens are still unset — the only way the direct
-  // `decimals0`/`decimals1` getters could be authoritative without
-  // a fallback is on FPMMs, which always have tokens set from their
-  // factory event. If tokens aren't set yet, this is a pre-start_block
-  // VP whose `selfHealWrappedExchangeId` couldn't backfill from
-  // `BiPoolExchange` (transient seed failure) — running the direct
-  // RPC here without a fallback is a wasted call. The next event,
-  // post-reverse-link backfill, will retry with tokens populated.
-  if (!pool.token0 || !pool.token1) return pool;
-  const poolAddr = extractAddressFromPoolId(pool.id);
-  const [dec0Raw, dec1Raw] = await Promise.all([
-    context.effect(tokenDecimalsScalingEffect, {
-      chainId: pool.chainId,
-      poolAddress: poolAddr,
-      fn: "decimals0",
-      fallbackTokenAddress: pool.token0,
-    }),
-    context.effect(tokenDecimalsScalingEffect, {
-      chainId: pool.chainId,
-      poolAddress: poolAddr,
-      fn: "decimals1",
-      fallbackTokenAddress: pool.token1,
-    }),
-  ]);
-  const parsed = parseDecimalsPair(dec0Raw, dec1Raw);
-  if (!parsed.tokenDecimalsKnown) return pool;
-  return { ...pool, ...parsed };
-}
-
-/** Self-heal `rebalanceThresholdAbove/Below` when the factory's
- * `rebalanceThresholdsEffect` failed at deploy. Without this, a transient
- * RPC blip would permanently leave both split fields at 0 → derive returns
- * null forever → the entity-derived path is dead for that pool. Block-
- * scoped read (effect is `cache: false` because thresholds are governance-
- * mutable). The caller's own Pool.set persists the healed values. */
-export async function selfHealRebalanceThresholds(
-  context: { effect: EffectCaller },
-  pool: Pool,
-  blockNumber: bigint,
-): Promise<Pool> {
-  if (
-    pool.rebalanceThresholdsKnown ||
-    pool.source === "" ||
-    isVirtualPool(pool)
-  ) {
-    return pool;
-  }
-  const poolAddr = extractAddressFromPoolId(pool.id);
-  const thresholds = await context.effect(rebalanceThresholdsEffect, {
-    chainId: pool.chainId,
-    poolAddress: poolAddr,
-    blockNumber,
-  });
-  if (thresholds === undefined) return pool;
-  // Refresh the legacy `rebalanceThreshold` only when at least one side is
-  // configured. Both-zero means "never rebalance" — leave the legacy field
-  // at whatever the next state-sync event pins.
-  const broadest = Math.max(thresholds.above, thresholds.below);
-  return {
-    ...pool,
-    rebalanceThresholdAbove: thresholds.above,
-    rebalanceThresholdBelow: thresholds.below,
-    rebalanceThresholdsKnown: true,
-    rebalanceThreshold: broadest > 0 ? broadest : pool.rebalanceThreshold,
-  };
-}
-
-/** Self-heal `Pool.wrappedExchangeId` for VirtualPools whose
- * `VirtualPoolDeployed` event fired pre-start_block (and so the bytecode
- * extraction in the factory handler never ran).
- *
- * Authoritative VP detector is `vpExchangeIdEffect` (bytecode pattern).
- * The previous source-based gate (`isVirtualPool(pool)`) read
- * `pool.source.includes("virtual")` — but VirtualPool.Swap / Mint / Burn
- * handlers reuse the `fpmm_*` source keys (intentional: they share
- * priority with FPMM events for `pickPreferredSource`), so a pre-
- * start_block VP whose first observed event was one of those would
- * never get the "virtual" substring and the heal would be skipped.
- * The bytecode pattern is definitive — it returns null for FPMMs and
- * `{exchangeProvider, exchangeId}` for VPs.
- *
- * The effect is `cache: true` and bytecode is immutable, so the RPC
- * fires once per address across the whole sync regardless of whether
- * the address is a VP or an FPMM. `vpExchangeIdEffect` distinguishes
- * "got bytecode, not a VP" (cached as a permanent miss) from "RPC
- * threw" (transient, not cached) — see the effect's comment for the
- * discriminator. After the cache populates, FPMMs cost zero RPC on
- * subsequent events.
- *
- * Also patches the matching `BiPoolExchange.wrappedByPoolId` back-
- * reference if the exchange row already exists, so the dashboard's
- * `wrappedByPoolId`-keyed GraphQL query finds the join after a heal
- * that only the Pool side knew about. When the BiPoolExchange row is
- * present, also backfills `token0`/`token1` from its `asset0`/`asset1`
- * if the Pool was created without them (Swap/Mint/Burn-first scenario:
- * those events don't carry the pair tokens, so the first VP swap
- * would otherwise show "?" symbols + zero USD valuation until some
- * later asset-bearing event lands).
- *
- * No-op on pools that already have `wrappedExchangeId` set. */
-export async function selfHealWrappedExchangeId(
-  context: PoolContext,
-  pool: Pool,
-  blockNumber: bigint,
-  blockTimestamp: bigint,
-): Promise<Pool> {
-  // Fully-healed gate: VP-confirmed AND tokens populated AND BiPoolExchange
-  // row seeded. Token presence and exchange presence both block retries
-  // on transient `poolExchangeEffect` / `tokenDecimalsScalingEffect`
-  // failures. Without the exchange-row check, a `VirtualPoolDeployed`
-  // event that succeeds in setting tokens but fails the exchange seed
-  // RPC would leave `wrappedExchangeId` pinned + tokens populated → the
-  // gate would skip heal forever, leaving no `BiPoolExchange` row and
-  // an empty `referenceRateFeedID` (round 7 codex #3). The bytecode
-  // effect is `cache:true`, so re-running the heal on every event for
-  // a fully-healed pool is cheap.
-  if (pool.wrappedExchangeId && pool.token0 && pool.token1) {
-    const exchangeRowId = `${pool.chainId}-${pool.wrappedExchangeId}`;
-    const existing = await context.BiPoolExchange.get(exchangeRowId);
-    if (existing?.wrappedByPoolId === pool.id) return pool;
-    // Else fall through — exchange seed still owed.
-    // If the row exists but is missing the back-reference, keep healing:
-    // older exchange-first rows can be marked checked before the VP link
-    // becomes visible, and the dashboard's VP exchange query keys on
-    // `wrappedByPoolId`.
-  }
-  const poolAddr = extractAddressFromPoolId(pool.id);
-  const result = await context.effect(vpExchangeIdEffect, {
-    chainId: pool.chainId,
-    vpAddress: poolAddr,
-  });
-  if (!result) return pool;
-  const exchangeRowId = `${pool.chainId}-${result.exchangeId}`;
-  let exchange = await context.BiPoolExchange.get(exchangeRowId);
-  // Round 3 #5: seed the BiPoolExchange row inline if it doesn't exist
-  // yet. Pre-start_block ordering: VP.Swap is the first observed event,
-  // BiPoolManager.ExchangeCreated fired before our start_block (so it
-  // never replays). Without seeding here, token + decimal backfill
-  // can't run, and the swap handler immediately persists
-  // `SwapEvent.volumeUsdWei` from default 18/18 decimals → mis-scaled
-  // forever (the later reverse-link backfill only updates the Pool
-  // row, not historical SwapEvent / leaderboard rows). RPC-fetch the
-  // struct via the same effect `BiPoolManager.ExchangeCreated` uses
-  // and materialize a row keyed on the bytecode-extracted
-  // exchangeProvider/exchangeId. Skip on RPC failure (transient) and
-  // on all-zero struct (destroyed exchange).
-  if (!exchange) {
-    const struct = await context.effect(poolExchangeEffect, {
-      chainId: pool.chainId,
-      exchangeProvider: result.exchangeProvider,
-      exchangeId: result.exchangeId,
-      blockNumber,
-    });
-    // Match `ensureBiPoolExchange`'s "destroyed exchange" test: reject
-    // ONLY the all-zero-and-no-pricing-module case. Pre-`BucketsUpdated`
-    // (newly-created exchange before its first bucket reset) returns
-    // zero buckets but has a real `pricingModule` + assets — that's a
-    // valid seed. A bucket-zero gate would reject those and skip the
-    // token/decimal backfill on the first VP swap.
-    if (struct && struct.pricingModule !== ZERO_ADDRESS) {
-      const seeded = {
-        id: exchangeRowId,
-        chainId: pool.chainId,
-        exchangeId: result.exchangeId,
-        exchangeProvider: result.exchangeProvider,
-        asset0: struct.asset0,
-        asset1: struct.asset1,
-        pricingModule: struct.pricingModule,
-        pricingModuleName:
-          lookupPricingModuleName(pool.chainId, struct.pricingModule) ??
-          undefined,
-        spread: struct.spread,
-        referenceRateFeedID: struct.referenceRateFeedID,
-        referenceRateResetFrequency: struct.referenceRateResetFrequency,
-        minimumReports: struct.minimumReports,
-        stablePoolResetSize: struct.stablePoolResetSize,
-        bucket0: struct.bucket0,
-        bucket1: struct.bucket1,
-        lastBucketUpdate: struct.lastBucketUpdate,
-        isDeprecated: false,
-        wrappedByPoolId: pool.id,
-        wrappedByPoolIdChecked: true,
-        createdAtBlock: blockNumber,
-        createdAtTimestamp: blockTimestamp,
-        updatedAtBlock: blockNumber,
-        updatedAtTimestamp: blockTimestamp,
-      };
-      context.BiPoolExchange.set(seeded);
-      exchange = seeded;
-    }
-  }
-  let healedFeedId = pool.referenceRateFeedID;
-  let healedToken0 = pool.token0;
-  let healedToken1 = pool.token1;
-  let healedToken0Decimals = pool.token0Decimals;
-  let healedToken1Decimals = pool.token1Decimals;
-  // Track per-leg whether THIS pass actually fetched decimals (vs.
-  // inheriting them from the spread). Round 5 used `dec > 0` which
-  // was satisfied by the schema-default 18 — codex caught this as a
-  // false-positive that would prevent `selfHealTokenDecimals` from
-  // ever retrying a transient blip on a non-18dp leg. Only flip
-  // `tokenDecimalsKnown=true` when both legs were fetched this pass.
-  let fetchedDec0 = false;
-  let fetchedDec1 = false;
-  if (exchange) {
-    if (exchange.wrappedByPoolId !== pool.id) {
-      context.BiPoolExchange.set({
-        ...exchange,
-        wrappedByPoolId: pool.id,
-        wrappedByPoolIdChecked: true,
-        updatedAtBlock: blockNumber,
-        updatedAtTimestamp: blockTimestamp,
-      });
-    }
-    // Once we've set wrappedByPoolId, ensureBiPoolExchange's null-guarded
-    // retry path won't re-run the feedID mirror on subsequent BucketsUpdated
-    // events. Mirror it here too so the oracle-price tile lights up
-    // immediately on first self-heal — otherwise it'd stay "—" until the
-    // NEXT bucket reset (~360s), even though the feedID is in scope.
-    await mirrorFeedIdToPool(
-      context,
-      pool.id,
-      exchange.referenceRateFeedID,
-      blockNumber,
-      blockTimestamp,
-    );
-    // CRUCIAL: also flow the mirrored feedID back to the caller so
-    // `upsertPool`'s next spread doesn't overwrite the just-persisted
-    // value. `mirrorFeedIdToPool` does its own `Pool.set`, but the
-    // healed Pool returned from this function feeds into `next` in
-    // upsertPool which then persists again — without carrying the
-    // updated feedID here, that second write would blank the mirror
-    // that just landed.
-    if (
-      exchange.referenceRateFeedID &&
-      exchange.referenceRateFeedID !== ZERO_ADDRESS
-    ) {
-      healedFeedId = exchange.referenceRateFeedID;
-    }
-    // Backfill pair tokens. Swap/Mint/Burn-first scenario: the Pool was
-    // created via getOrCreate without `defaults.token0/token1` (those
-    // events don't carry the pair). Without this, the first swap is
-    // valued at $0 and the dashboard renders "?" symbols until some
-    // later asset-bearing event arrives — which may never happen for
-    // pre-start_block VPs. The BiPoolExchange row already has the
-    // assets from BiPoolManager.ExchangeCreated. Only fill on miss
-    // (existing token0/token1 wins on direct conflict). Skip
-    // ZERO_ADDRESS — a zeroed exchange row (transient RPC backfill at
-    // ExchangeCreated time) shouldn't pin token0/token1 to 0x0 + then
-    // fire an `eth_call decimals()` against it. Mirrors the same guard
-    // in `mirrorTokensAndDecimalsToPool` (reverse-link).
-    // Backfill pair tokens + decimals AS A UNIT (round 4 codex #1).
-    // Pinning the address while leaving decimals at the default 18 was
-    // a bug: the new gate above (`&& pool.token0 && pool.token1`) would
-    // see token0 set + skip retries, leaving a 6dp USDC leg permanently
-    // mis-scaled. Fix: only pin the token when decimals fetch succeeds.
-    // On transient `tokenDecimalsScalingEffect` failure, leave both
-    // unset — gate stays open, next event retries the pair.
-    //
-    // Default 18 is wrong for any non-18dp leg (e.g. USDC at 6dp) —
-    // `buildSwapTraderFields` runs immediately after `upsertPool`
-    // returns and uses these to scale `volumeUsdWei`. `tokenDecimalsScalingEffect`
-    // is cache:true (one RPC per chain/token per sync) and tries
-    // `decimals0()`/`decimals1()` on the pool first (null for VPs — that
-    // getter lives on FPMM), then falls back to ERC20 `decimals()` on
-    // the token address. ZERO_ADDRESS guard mirrors the same in
-    // `mirrorTokensAndDecimalsToPool`.
-    const fillToken0 =
-      !healedToken0 && exchange.asset0 && exchange.asset0 !== ZERO_ADDRESS;
-    const fillToken1 =
-      !healedToken1 && exchange.asset1 && exchange.asset1 !== ZERO_ADDRESS;
-    if (fillToken0) {
-      const dec = await context.effect(tokenDecimalsScalingEffect, {
-        chainId: pool.chainId,
-        poolAddress: poolAddr,
-        fn: "decimals0",
-        fallbackTokenAddress: exchange.asset0,
-      });
-      if (dec) {
-        healedToken0 = exchange.asset0;
-        healedToken0Decimals = scalingFactorToDecimals(dec) ?? 18;
-        fetchedDec0 = true;
-      }
-    }
-    if (fillToken1) {
-      const dec = await context.effect(tokenDecimalsScalingEffect, {
-        chainId: pool.chainId,
-        poolAddress: poolAddr,
-        fn: "decimals1",
-        fallbackTokenAddress: exchange.asset1,
-      });
-      if (dec) {
-        healedToken1 = exchange.asset1;
-        healedToken1Decimals = scalingFactorToDecimals(dec) ?? 18;
-        fetchedDec1 = true;
-      }
-    }
-  }
-  // Flip `tokenDecimalsKnown` only when this pass actually fetched both
-  // legs' decimals via the effect. Inheriting `decimals > 0` from the
-  // schema default (18) is NOT proof of trust — that's the
-  // false-positive codex flagged. If decimals were already known
-  // before this call, preserve true. Otherwise the cross-pass
-  // coordination falls to `selfHealTokenDecimals` (which fires both
-  // legs independently and sets the flag when both succeed).
-  const decimalsKnown = pool.tokenDecimalsKnown || (fetchedDec0 && fetchedDec1);
-  return {
-    ...pool,
-    wrappedExchangeId: result.exchangeId,
-    referenceRateFeedID: healedFeedId,
-    token0: healedToken0,
-    token1: healedToken1,
-    token0Decimals: healedToken0Decimals,
-    token1Decimals: healedToken1Decimals,
-    tokenDecimalsKnown: decimalsKnown,
-  };
-}
-
-/** Mirror a v2 exchange's `referenceRateFeedID` onto its wrapping
- * VirtualPool's Pool row so the existing SortedOracles `getPoolsByFeed`
- * lookup picks up the VP naturally. Idempotent — no-op when already in
- * sync. Skips zero/empty feedIDs (still pre-RPC-backfill or destroyed
- * exchange) so a transient source value can't blank a previously-set
- * link. Used from both directions of the VP↔BiPoolExchange linkage
- * (VirtualPoolDeployed forward, BiPoolManager.ExchangeCreated reverse,
- * plus `selfHealWrappedExchangeId` post-self-heal).
- *
- * Context narrowed to just `Pool` access — callers from both PoolContext
- * (upsertPool path) and HandlerContext (event handlers) work. */
-export async function mirrorFeedIdToPool(
-  context: { Pool: PoolContext["Pool"] },
-  poolId: string,
-  feedId: string,
-  blockNumber: bigint,
-  blockTimestamp: bigint,
-): Promise<void> {
-  if (!feedId || feedId === ZERO_ADDRESS) return;
-  const pool = await context.Pool.get(poolId);
-  if (!pool || pool.referenceRateFeedID === feedId) return;
-  context.Pool.set({
-    ...pool,
-    referenceRateFeedID: feedId,
-    updatedAtBlock: blockNumber,
-    updatedAtTimestamp: blockTimestamp,
-  });
-}
-
-/** Reverse-link backfill: when a `BiPoolExchange` row is created/updated
- * AFTER its wrapping VP has self-healed (so the back-link
- * `wrappedByPoolId` is already known), mirror `asset0`/`asset1` and the
- * matching decimals onto the Pool row. Mirrors the forward backfill in
- * `selfHealWrappedExchangeId` for the heal-before-exchange ordering.
- *
- * Without this, a VP that healed when no `BiPoolExchange` row existed
- * yet keeps `token0`/`token1` empty + decimals at the 18/18 default
- * forever — the heal helper only mirrored fields it could see at the
- * time, and `mirrorFeedIdToPool` only handles the feedID. The first
- * VP swap then valuates at `?/?` with mis-scaled USD until some later
- * asset-bearing event lands (which may never happen for pre-start_block
- * VPs).
- *
- * Idempotent: only fills missing fields (existing token0/token1 wins on
- * direct conflict). Skips the decimals RPC when the address backfill
- * isn't needed. Wider context type (handler `effect` + `Pool` access)
- * because the decimals fetcher uses `tokenDecimalsScalingEffect`. */
-export async function mirrorTokensAndDecimalsToPool(
-  context: {
-    Pool: PoolContext["Pool"];
-    effect: PoolContext["effect"];
-  },
-  poolId: string,
-  asset0: string,
-  asset1: string,
-  blockNumber: bigint,
-  blockTimestamp: bigint,
-): Promise<void> {
-  const pool = await context.Pool.get(poolId);
-  if (!pool) return;
-  // Fully verified: tokens populated AND `tokenDecimalsKnown=true`.
-  // Otherwise fall through so cross-pass / partial-fetch state can be
-  // re-validated (round 7 codex #1+#2).
-  if (pool.tokenDecimalsKnown && pool.token0 && pool.token1) return;
-  const fillToken0 = !pool.token0 && asset0 && asset0 !== ZERO_ADDRESS;
-  const fillToken1 = !pool.token1 && asset1 && asset1 !== ZERO_ADDRESS;
-  const poolAddr = extractAddressFromPoolId(poolId);
-  // Round 4 codex #1: pin token + decimals as a unit. If the decimals
-  // fetch transiently fails, leave the address unset too so the next
-  // event re-tries the pair — pinning the address while leaving
-  // decimals at default 18 would lock in a mis-scaled valuation for
-  // any non-18dp token (e.g. USDC at 6dp).
-  let nextToken0 = pool.token0;
-  let nextToken1 = pool.token1;
-  let nextToken0Decimals = pool.token0Decimals;
-  let nextToken1Decimals = pool.token1Decimals;
-  // Per-leg track whether THIS pass actually fetched decimals. The flag
-  // flips below only when BOTH legs end up known — either via a fresh
-  // fetch this pass OR via a fetch in this pass that confirms the
-  // existing value (cross-pass case). `tokenDecimalsScalingEffect` is
-  // `cache:true`, so re-fetching a leg whose decimals are already in
-  // place is essentially free.
-  let fetchedDec0 = false;
-  let fetchedDec1 = false;
-  // Round 7 codex #1+#2: even when both tokens are populated already,
-  // run the decimals fetch when `tokenDecimalsKnown=false` so the flag
-  // gets flipped on the cross-pass case (each leg landed via a
-  // separate event). Use the existing token address as the fallback
-  // when we're not filling.
-  const fallback0 = fillToken0 ? asset0 : pool.token0;
-  const fallback1 = fillToken1 ? asset1 : pool.token1;
-  if (fallback0 && fallback0 !== ZERO_ADDRESS) {
-    const dec = await context.effect(tokenDecimalsScalingEffect, {
-      chainId: pool.chainId,
-      poolAddress: poolAddr,
-      fn: "decimals0",
-      fallbackTokenAddress: fallback0,
-    });
-    if (dec) {
-      if (fillToken0) nextToken0 = asset0;
-      nextToken0Decimals = scalingFactorToDecimals(dec) ?? 18;
-      fetchedDec0 = true;
-    }
-  }
-  if (fallback1 && fallback1 !== ZERO_ADDRESS) {
-    const dec = await context.effect(tokenDecimalsScalingEffect, {
-      chainId: pool.chainId,
-      poolAddress: poolAddr,
-      fn: "decimals1",
-      fallbackTokenAddress: fallback1,
-    });
-    if (dec) {
-      if (fillToken1) nextToken1 = asset1;
-      nextToken1Decimals = scalingFactorToDecimals(dec) ?? 18;
-      fetchedDec1 = true;
-    }
-  }
-  // Decide if any state actually changed. A change is: a token address
-  // newly filled, decimals updated, or the trust-flag flipping.
-  const decimalsKnownNext =
-    pool.tokenDecimalsKnown || (fetchedDec0 && fetchedDec1);
-  const tokensChanged =
-    nextToken0 !== pool.token0 || nextToken1 !== pool.token1;
-  const decimalsChanged =
-    nextToken0Decimals !== pool.token0Decimals ||
-    nextToken1Decimals !== pool.token1Decimals;
-  const flagFlipped = decimalsKnownNext !== pool.tokenDecimalsKnown;
-  if (!tokensChanged && !decimalsChanged && !flagFlipped) return;
-  context.Pool.set({
-    ...pool,
-    token0: nextToken0,
-    token1: nextToken1,
-    token0Decimals: nextToken0Decimals,
-    token1Decimals: nextToken1Decimals,
-    tokenDecimalsKnown: decimalsKnownNext,
-    updatedAtBlock: blockNumber,
-    updatedAtTimestamp: blockTimestamp,
-  });
-}
-
-export type SnapshotContext = {
-  PoolSnapshot: {
-    get: (id: string) => Promise<PoolSnapshot | undefined>;
-    set: (entity: PoolSnapshot) => void;
-  };
-  PoolDailySnapshot: {
-    get: (id: string) => Promise<PoolDailySnapshot | undefined>;
-    set: (entity: PoolDailySnapshot) => void;
-  };
-};
 
 /** Default oracle field values (for VirtualPools or when RPC call fails).
  *
@@ -1413,154 +523,4 @@ export const upsertPool = async ({
 
   context.Pool.set(final);
   return final;
-};
-
-// ---------------------------------------------------------------------------
-// PoolSnapshot upsert
-// ---------------------------------------------------------------------------
-
-export const upsertSnapshot = async ({
-  context,
-  pool,
-  blockTimestamp,
-  blockNumber,
-  swapDelta,
-  rebalanceDelta,
-  mintDelta,
-  burnDelta,
-}: {
-  context: SnapshotContext;
-  pool: Pool;
-  blockTimestamp: bigint;
-  blockNumber: bigint;
-  swapDelta?: { volume0: bigint; volume1: bigint };
-  rebalanceDelta?: boolean;
-  mintDelta?: boolean;
-  burnDelta?: boolean;
-}): Promise<void> => {
-  const hourTs = hourBucket(blockTimestamp);
-  const id = snapshotId(pool.id, hourTs);
-  const existing = await context.PoolSnapshot.get(id);
-
-  const snapshot: PoolSnapshot = existing
-    ? {
-        ...existing,
-        reserves0: pool.reserves0,
-        reserves1: pool.reserves1,
-        swapCount: existing.swapCount + (swapDelta ? 1 : 0),
-        swapVolume0: existing.swapVolume0 + (swapDelta?.volume0 ?? 0n),
-        swapVolume1: existing.swapVolume1 + (swapDelta?.volume1 ?? 0n),
-        rebalanceCount: existing.rebalanceCount + (rebalanceDelta ? 1 : 0),
-        mintCount: existing.mintCount + (mintDelta ? 1 : 0),
-        burnCount: existing.burnCount + (burnDelta ? 1 : 0),
-        cumulativeSwapCount: pool.swapCount,
-        cumulativeVolume0: pool.notionalVolume0,
-        cumulativeVolume1: pool.notionalVolume1,
-        blockNumber,
-      }
-    : {
-        id,
-        chainId: pool.chainId,
-        poolId: pool.id,
-        timestamp: hourTs,
-        reserves0: pool.reserves0,
-        reserves1: pool.reserves1,
-        swapCount: swapDelta ? 1 : 0,
-        swapVolume0: swapDelta?.volume0 ?? 0n,
-        swapVolume1: swapDelta?.volume1 ?? 0n,
-        rebalanceCount: rebalanceDelta ? 1 : 0,
-        mintCount: mintDelta ? 1 : 0,
-        burnCount: burnDelta ? 1 : 0,
-        cumulativeSwapCount: pool.swapCount,
-        cumulativeVolume0: pool.notionalVolume0,
-        cumulativeVolume1: pool.notionalVolume1,
-        blockNumber,
-      };
-
-  context.PoolSnapshot.set(snapshot);
-
-  // Also write the day-bucketed rollup. Callers never need to invoke this
-  // directly — handlers call upsertSnapshot, both entities get updated.
-  await upsertDailySnapshot({
-    context,
-    pool,
-    blockTimestamp,
-    blockNumber,
-    swapDelta,
-    rebalanceDelta,
-    mintDelta,
-    burnDelta,
-  });
-};
-
-// ---------------------------------------------------------------------------
-// PoolDailySnapshot upsert — same read-merge-write pattern as upsertSnapshot,
-// but bucketed per UTC day. Lets full-history pool charts fit in a single
-// Hasura page (Envio's hosted endpoint caps every query at 1000 rows).
-// Invoked from upsertSnapshot; exported so tests can exercise it directly.
-// ---------------------------------------------------------------------------
-
-export const upsertDailySnapshot = async ({
-  context,
-  pool,
-  blockTimestamp,
-  blockNumber,
-  swapDelta,
-  rebalanceDelta,
-  mintDelta,
-  burnDelta,
-}: {
-  context: SnapshotContext;
-  pool: Pool;
-  blockTimestamp: bigint;
-  blockNumber: bigint;
-  swapDelta?: { volume0: bigint; volume1: bigint };
-  rebalanceDelta?: boolean;
-  mintDelta?: boolean;
-  burnDelta?: boolean;
-}): Promise<void> => {
-  const dayTs = dayBucket(blockTimestamp);
-  const id = dailySnapshotId(pool.id, dayTs);
-  const existing = await context.PoolDailySnapshot.get(id);
-
-  const snapshot: PoolDailySnapshot = existing
-    ? {
-        ...existing,
-        reserves0: pool.reserves0,
-        reserves1: pool.reserves1,
-        swapCount: existing.swapCount + (swapDelta ? 1 : 0),
-        swapVolume0: existing.swapVolume0 + (swapDelta?.volume0 ?? 0n),
-        swapVolume1: existing.swapVolume1 + (swapDelta?.volume1 ?? 0n),
-        rebalanceCount: existing.rebalanceCount + (rebalanceDelta ? 1 : 0),
-        mintCount: existing.mintCount + (mintDelta ? 1 : 0),
-        burnCount: existing.burnCount + (burnDelta ? 1 : 0),
-        cumulativeSwapCount: pool.swapCount,
-        cumulativeVolume0: pool.notionalVolume0,
-        cumulativeVolume1: pool.notionalVolume1,
-        cumulativeHealthBinarySeconds: pool.healthBinarySeconds,
-        cumulativeHealthTotalSeconds: pool.healthTotalSeconds,
-        blockNumber,
-      }
-    : {
-        id,
-        chainId: pool.chainId,
-        poolId: pool.id,
-        timestamp: dayTs,
-        reserves0: pool.reserves0,
-        reserves1: pool.reserves1,
-        swapCount: swapDelta ? 1 : 0,
-        swapVolume0: swapDelta?.volume0 ?? 0n,
-        swapVolume1: swapDelta?.volume1 ?? 0n,
-        rebalanceCount: rebalanceDelta ? 1 : 0,
-        mintCount: mintDelta ? 1 : 0,
-        burnCount: burnDelta ? 1 : 0,
-        cumulativeSwapCount: pool.swapCount,
-        cumulativeVolume0: pool.notionalVolume0,
-        cumulativeVolume1: pool.notionalVolume1,
-        cumulativeHealthBinarySeconds: pool.healthBinarySeconds,
-        cumulativeHealthTotalSeconds: pool.healthTotalSeconds,
-        blockNumber,
-      };
-
-  context.PoolDailySnapshot.set(snapshot);
 };
